Extract DivIconMarker to separate file
